Step 1 of 2 in fixing Sendwithus folder caps
Renamed the Sendwithus folders to the correct capitalization, but with
an _temp suffix.  They were named correctly in Visual Studio and my
local repo, but weren't updated on GitHub.  Without the _temp suffix,
they git doesn't register the directory name change because it's the
same name as before but with different capitalization.