Removed any order dependencies in unit tests
Some Delete and Update unit tests would only work if another test had
already been run to create a new resource of the type that was to be
deleted or updated.  Changed them to add a new resource from within
those Delete and Update tests so there's no longer a dependency on the
order of the tests being run.
-Minor formatting changes to Template.cs comments.